Huge progress with JS scripts

+++ Tools/CordinatestConvert/Country.cs
@@ -3,7 +3,7 @@ namespace CoordinatesConvertor
     public class Country
     {
         public string name;
-        public float[][][] coordinates;
+        public double[][][] coordinates;
     }
